requirements.txt
elasticsearch==7.17.7
farm-haystack @ git+https://github.com/deepset-ai/haystack.git@8db7dfb8841ba0184b4a46ce75601dad8dddddf6
Flask==2.2.2
flask_bootstrap
flask_wtf
wtforms
numpy==1.23.4
praw==7.6.0